+++ dotnet/test/UShell.ServerCommands.DemoWebService/Startup.cs
@@ -77,6 +77,8 @@ namespace Demo {
       );
 
       //////////////////////////////////////////////////////////////////////////////////////////
+      
+      services.AddUjmwStandardSwaggerGen();
     }
 
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.