+++ doc/changelog.md
@@ -8,7 +8,7 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
+* Completed implementation of IHostedService, (**MVP**-state is now reached)